+++ cocos/3d/CCSkeleton3D.cpp
@@ -303,9 +303,9 @@ Bone3D* Skeleton3D::getRootBone(int index) const
 
 int Skeleton3D::getBoneIndex(Bone3D* bone) const
 {
-    for (int i = 0, size = _bones.size(); i < size; ++i) {
+    for (ssize_t i = 0, size = _bones.size(); i < size; ++i) {
         if (_bones.at(i) == bone)
-            return i;
+            return static_cast<int>(i);
     }
 
     return -1;
